+++ loader/schemas/show_user/show_position_frame/simulation/land_simulation.py
@@ -1,11 +1,12 @@
 from typing import TYPE_CHECKING
 
+import numpy as np
+
 from loader.parameters import FRAME_PARAMETERS, LAND_PARAMETERS
 
 from .position_simulation import SimulationInfo, linear_interpolation
 
 if TYPE_CHECKING:
-    import numpy as np
     from numpy.typing import NDArray
 
 
@@ -57,6 +58,43 @@ def land_simulation(
     ) + generate_land_second_part(
         land_start_position,
     )
+    return [
+        SimulationInfo(
+            frame=frame_begin + frame_index,
+            position=land_position,
+        )
+        for frame_index, land_position in enumerate(land_positions)
+    ]
+
+
+def rtl_simulation(
+    takeoff_position: tuple[float, float, float],
+    rtl_start_position: tuple[float, float, float],
+    frame_begin: int,
+    speed: float = 2.0,
+) -> list["SimulationInfo"]:
+    rtl_end_position = (
+        takeoff_position[0],
+        takeoff_position[1],
+        0.0,
+    )
+
+    rtl_duration = round(
+        np.linalg.norm(
+            np.array(rtl_end_position, dtype=np.float32)
+            - np.array(rtl_start_position, dtype=np.float32)
+        )
+        * 1.57
+        / speed
+        * FRAME_PARAMETERS.fps
+    )
+
+    land_positions = linear_interpolation(
+        rtl_start_position,
+        rtl_end_position,
+        rtl_duration,
+    )
+
     return [
         SimulationInfo(
             frame=frame_begin + frame_index,