+++ tests/schemas/migrations/test_ijg_to_su.py
@@ -6,6 +6,7 @@ from loader.parameters import (
     IOSTAR_PHYSIC_PARAMETERS_MAX,
     IOSTAR_PHYSIC_PARAMETERS_RECOMMENDATION,
     LandType,
+    MagicNumber,
 )
 from loader.schemas import IostarJsonGcs, ShowUser
 from loader.schemas.matrix import get_matrix
@@ -16,6 +17,7 @@ from tests.strategies import (
     slow,
     st_angle_takeoff,
     st_land_type,
+    st_magic_number,
     st_matrix_with_shape,
     st_scale,
     st_step_takeoff,
@@ -29,6 +31,7 @@ if TYPE_CHECKING:
     matrix_with_shape=st_matrix_with_shape(),
     step_takeoff=st_step_takeoff,
     angle_takeoff=st_angle_takeoff,
+    magic_number=st_magic_number,
     scale=st_scale,
     land_type=st_land_type,
 )
@@ -37,6 +40,7 @@ def test_ijg_to_su(
     matrix_with_shape: tuple["NDArray[np.intp]", int, int, int],
     step_takeoff: float,
     angle_takeoff: float,
+    magic_number: MagicNumber,
     scale: int,
     land_type: LandType,
 ) -> None:
@@ -44,12 +48,15 @@ def test_ijg_to_su(
     show_user = get_valid_show_user(
         ShowUserConfiguration(
             matrix=matrix,
-            step=step_takeoff,
+            step_x=step_takeoff,
+            step_y=step_takeoff,
             angle_takeoff=angle_takeoff,
+            magic_number=magic_number,
         ),
     )
-    show_user.scale = scale
-    show_user.land_type = land_type
+    if magic_number >= MagicNumber.v3:  # pragma: no cover
+        show_user.scale = scale
+        show_user.land_type = land_type
     iostar_json_gcs = IostarJsonGcs.from_show_user(show_user)
     iostar_json_gcs_angle_takeoff_rad = -np.deg2rad(iostar_json_gcs.show.angle_takeoff)
     assert is_angles_equal(show_user.angle_takeoff, iostar_json_gcs_angle_takeoff_rad)
@@ -62,8 +69,12 @@ def test_ijg_to_su(
         export_import_autopilot_format.angle_takeoff,
         iostar_json_gcs_angle_takeoff_rad,
     )
-    assert show_user.nb_x == export_import_autopilot_format.nb_x == nb_x
-    assert show_user.nb_y == export_import_autopilot_format.nb_y == nb_y
+    assert (
+        show_user.nb_x == export_import_autopilot_format.nb_x == nb_x
+    ), f"{show_user.nb_x=} {export_import_autopilot_format.nb_x=} {nb_x=}"
+    assert (
+        show_user.nb_y == export_import_autopilot_format.nb_y == nb_y
+    ), f"{show_user.nb_y=} {export_import_autopilot_format.nb_y=} {nb_y=}"
     assert (
         show_user.nb_drones_per_family
         == export_import_autopilot_format.nb_drones_per_family
@@ -76,7 +87,8 @@ def test_show_user_from_iostar_json_gcs_without_physic_parameters() -> None:
     show_user = get_valid_show_user(
         ShowUserConfiguration(
             matrix=get_matrix(),
-            step=1,
+            step_x=1,
+            step_y=1,
             angle_takeoff=0,
         ),
     )
@@ -90,7 +102,8 @@ def test_show_user_from_iostar_json_gcs_with_physic_parameters() -> None:
     show_user = get_valid_show_user(
         ShowUserConfiguration(
             matrix=get_matrix(),
-            step=1,
+            step_x=1,
+            step_y=1,
             angle_takeoff=0,
         ),
     )
@@ -99,3 +112,13 @@ def test_show_user_from_iostar_json_gcs_with_physic_parameters() -> None:
     iostar_json_gcs.physic_parameters = None
     show_user.from_iostar_json_gcs(iostar_json_gcs)
     assert show_user.physic_parameters == IOSTAR_PHYSIC_PARAMETERS_MAX
+
+
+def test_import_show_with_step_format() -> None:
+    from pathlib import Path
+
+    dance_path = Path("iostar_json_gcs_valid_with_step_arg.json")
+    iostar_json_gcs = IostarJsonGcs.model_validate_json(dance_path.read_text())
+    show_user = ShowUser.from_iostar_json_gcs(iostar_json_gcs)
+
+    assert show_user.step_x == show_user.step_y